inbound: factor out helpers shared by fuzz and stack tests
Currently, the inbound HTTP fuzz logic and stack tests duplicate a fair
amount of test support code. This code is almost exactly the same
between the fuzz logic and tests, but it's duplicated rather than shared
--- introducing the opportunity for minor inconsistencies to sneak in,
especially because we don't actually test that the fuzz logic runs
successfully in CI. On the other hand, if we reused the same helpers as
the unit tests, we would at least be testing the support code on CI.
This would have prevented the fuzz breakage fixed in #1193, which was
caused by an inconsistency between the fuzz logic and stack test
versions of the `Target` type.

This PR factors out the duplicated code into a `test_util` module that's
enabled by both `cfg(test)` and `cfg(fuzzing)`. Both modules now use the
shared version of the test support code defined in that module.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>